+++ src/main/java/com/c2psi/businessmanagement/dtos/client/client/ClientDamageOperationDto.java
@@ -11,6 +11,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -24,7 +25,7 @@ public class ClientDamageOperationDto {
     OperationDto cltdoOperationDto;
     @NotNull(message = "The number in mvt cannot be null")
     @Positive(message = "The number in mvt must be positive")
-    Integer cltdoNumberinmvt;
+    BigDecimal cltdoNumberinmvt;
     /******************************
      * Relation between entities  *
      * ****************************/
