+++ src/main/java/com/c2psi/businessmanagement/dtos/client/client/ClientPackagingAccountDto.java
@@ -7,13 +7,14 @@ import lombok.Builder;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 public class ClientPackagingAccountDto {
     Long id;
     @NotNull(message = "The number of packaging in the account cannot be null")
-    Integer cpaNumber;
+    BigDecimal cpaNumber;
     /******************************
      * Relation between entities  *
      * ****************************/
