+++ src/main/java/com/c2psi/businessmanagement/dtos/pos/pos/PosPackagingOperationDto.java
@@ -9,6 +9,7 @@ import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -18,7 +19,7 @@ public class PosPackagingOperationDto {
     OperationDto pospoOperationDto;
     @NotNull(message = "The number in mvt cannot be null")
     @Positive(message = "The number in mvt must be positive")
-    Integer pospoNumberinmvt;
+    BigDecimal pospoNumberinmvt;
     /******************************
      * Relation between entities  *
      * ****************************/
