+++ src/main/java/com/c2psi/businessmanagement/dtos/stock/provider/ProviderCapsuleOperationDto.java
@@ -8,6 +8,7 @@ import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Positive;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -17,7 +18,7 @@ public class ProviderCapsuleOperationDto {
     OperationDto proscoOperationDto;
     @NotNull(message = "The number in mvt in the operation cannot be null")
     @Positive(message = "The number in mvt in the operation must be positive")
-    Integer procsoNumberinmvt;
+    BigDecimal procsoNumberinmvt;
     /******************************
      * Relation between entities  *
      * ****************************/
