+++ src/main/java/com/c2psi/businessmanagement/models/ClientDamageOperation.java
@@ -7,6 +7,7 @@ import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -27,7 +28,7 @@ public class ClientDamageOperation extends AbstractEntity {
     @Embedded
     Operation cltdoOperation;
 
-    Integer cltdoNumberinmvt;
+    BigDecimal cltdoNumberinmvt;
     /******************************
      * Relation between entities  *
      * ****************************/
