+++ src/main/java/com/c2psi/businessmanagement/models/ProviderCapsuleOperation.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Data
@@ -25,7 +26,7 @@ public class ProviderCapsuleOperation extends AbstractEntity {
     @Embedded
     Operation proscoOperation;
 
-    Integer procsoNumberinmvt;
+    BigDecimal procsoNumberinmvt;
     /******************************
      * Relation between entities  *
      * ****************************/
